fix(settings): inline History privacy UI and extract More browsers
Give Macroscope UI Consistency a real web-settings delta so retained
findings from e1f85f3 are re-evaluated against fixed primitives.
